move test to __init__.py no device_registry needed
